# Make B3 parser more readable.