@g-easy 2019-08-13
Benchmark before:
```
Benchmark                                Time             CPU   Iterations
--------------------------------------------------------------------------
BM_FromB3Headers_128bitTraceId        76.7 ns         76.7 ns      6863265
BM_FromB3Headers_64bitTraceId         45.2 ns         45.2 ns     15227647
BM_FromB3Headers_InvalidTraceId       25.1 ns         25.1 ns     30997309
```

After:
```
Benchmark                                Time             CPU   Iterations
--------------------------------------------------------------------------
BM_FromB3Headers_128bitTraceId        74.4 ns         74.4 ns      8511255
BM_FromB3Headers_64bitTraceId         43.1 ns         43.1 ns     16243200
BM_FromB3Headers_InvalidTraceId       29.0 ns         29.0 ns     23450057
```